fix(ci): split lib create-project lint paths
